drivers: thermal: call TSENS re-init only when register is ready to update
If TSENS controller reset pin is held asserted for long duration
and the same time if TSENS software detects TSENS reset and tries
to re-initialize TSENS controller, TSENS initialization may not
reflect in TSENS register due to reset pin is held asserted.

Invoke TSENS re-init routine only if software is able to write
TSENS register successfully. It makes sure that TSENS re-init
routine is able to re-initialize TSENS controller.

Add a check whether TSENS is re-enabled or not on post TSENS re-init.
If it is not re-enabled, re-try TSENS re-init routine.

Change-Id: Ia0d9b794b47854d269438ccfd9967cf7112aa5e5
Signed-off-by: Manaf Meethalavalappu Pallikunhi <manafm@codeaurora.org>
Signed-off-by: Gopala Krishna Nuthaki <gnuthaki@codeaurora.org>
Signed-off-by: DennySPb <dennyspb@gmail.com>